Implement Stripe webhook functionality and update environment variables
