+++ .env
@@ -5,4 +5,7 @@ VITE_SUPABASE_URL=https://hmrinsnimrwijlqtqahb.supabase.co
 
 # Stripe
 STRIPE_WEBHOOK_SECRET=whsec_vhmHpAO4qTVz0WvYufulJA9DBjRFLO3W
-STRIPE_SECRET_KEY=
+STRIPE_SECRET_KEY=sk_test_51J3
+
+# Server
+PORT=9099
